move color functions into utils
