+++ utils.go
@@ -1,8 +1,10 @@
 package main
 
 import (
+	"fmt"
 	"strings"
 
+	"github.com/fatih/color"
 	"github.com/jesseduffield/gocui"
 )
 
@@ -27,4 +29,14 @@ func withPadding(str string, padding int) string {
 		return str
 	}
 	return str + strings.Repeat(" ", padding-len(str))
+}
+
+func coloredString(str string, colorAttribute color.Attribute) string {
+	colour := color.New(colorAttribute)
+	return coloredStringDirect(str, colour)
+}
+
+// used for aggregating a few color attributes rather than just sending a single one
+func coloredStringDirect(str string, colour *color.Color) string {
+	return colour.SprintFunc()(fmt.Sprint(str))
 }
